aws/mercadolivre_deploy_20250923_162013/requirements_aws_clean.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==2.3.7
gunicorn==21.2.0
